backend/requirements.txt
# Requirements for Python backend
fastapi==0.109.0
uvicorn==0.27.0
yfinance==0.2.35
akshare==1.13.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
